Update README.md
In this project, I was presented with a corrupted group student picture. Its original color was changed, and the brightness and contrast was tampered as well.
And I was tasked to Use your Computer Vision image processing and restore the picture to its original color and contrast Write a code to show the pictures original color and correct the contrast and brightness.